do not use the Test suffix for non test classes
