+++ src/Symfony/Component/Form/Tests/VersionAwareTestTrait.php
@@ -14,7 +14,7 @@ namespace Symfony\Component\Form\Tests;
 /**
  * @deprecated since Symfony 7.2, use feature detection instead.
  */
-trait VersionAwareTest
+trait VersionAwareTestTrait
 {
     protected static int $supportedFeatureSetVersion = 404;
 
@@ -23,7 +23,7 @@ trait VersionAwareTest
      */
     protected function requiresFeatureSet(int $requiredFeatureSetVersion)
     {
-        trigger_deprecation('symfony/form', '7.2', 'The "%s" trait is deprecated, use feature detection instead.', VersionAwareTest::class);
+        trigger_deprecation('symfony/form', '7.2', 'The "%s" trait is deprecated, use feature detection instead.', VersionAwareTestTrait::class);
 
         if ($requiredFeatureSetVersion > static::$supportedFeatureSetVersion) {
             $this->markTestSkipped(\sprintf('Test requires features from symfony/form %.2f but only version %.2f is supported.', $requiredFeatureSetVersion / 100, static::$supportedFeatureSetVersion / 100));
